chore: split node.go into smaller files
