+++ node.go
@@ -18,10 +18,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"log/slog"
 	"net"
 	"net/http"
-	"runtime/debug"
 	"strconv"
 	"sync"
 	"time"
@@ -44,8 +42,6 @@ import (
 	"github.com/blinklabs-io/dingo/peergov"
 	"github.com/blinklabs-io/dingo/utxorpc"
 	ouroboros "github.com/blinklabs-io/gouroboros"
-	gledger "github.com/blinklabs-io/gouroboros/ledger"
-	lcommon "github.com/blinklabs-io/gouroboros/ledger/common"
 )
 
 type Node struct {
@@ -75,288 +71,6 @@ type Node struct {
 	shutdownOnce                     sync.Once
 	chainsyncIngressEligibilityMu    sync.RWMutex
 	chainsyncIngressEligibilityCache map[ouroboros.ConnectionId]bool
-}
-
-func plateauThreshold(stallTimeout time.Duration) time.Duration {
-	return max(2*stallTimeout, 4*time.Minute)
-}
-
-func (n *Node) isChainsyncIngressEligible(
-	connId ouroboros.ConnectionId,
-) bool {
-	if n.peerGov != nil {
-		return n.peerGov.IsChainSelectionEligible(connId)
-	}
-	n.chainsyncIngressEligibilityMu.RLock()
-	defer n.chainsyncIngressEligibilityMu.RUnlock()
-	if n.chainsyncIngressEligibilityCache == nil {
-		return false
-	}
-	eligible, ok := n.chainsyncIngressEligibilityCache[connId]
-	if !ok {
-		return false
-	}
-	return eligible
-}
-
-func (n *Node) setChainsyncIngressEligibility(
-	connId ouroboros.ConnectionId,
-	eligible bool,
-) {
-	n.chainsyncIngressEligibilityMu.Lock()
-	defer n.chainsyncIngressEligibilityMu.Unlock()
-	if n.chainsyncIngressEligibilityCache == nil {
-		n.chainsyncIngressEligibilityCache = make(
-			map[ouroboros.ConnectionId]bool,
-		)
-	}
-	n.chainsyncIngressEligibilityCache[connId] = eligible
-}
-
-func (n *Node) deleteChainsyncIngressEligibility(
-	connId ouroboros.ConnectionId,
-) {
-	n.chainsyncIngressEligibilityMu.Lock()
-	defer n.chainsyncIngressEligibilityMu.Unlock()
-	if n.chainsyncIngressEligibilityCache == nil {
-		return
-	}
-	delete(n.chainsyncIngressEligibilityCache, connId)
-}
-
-func (n *Node) handlePeerEligibilityChangedEvent(evt event.Event) {
-	e, ok := evt.Data.(peergov.PeerEligibilityChangedEvent)
-	if !ok {
-		return
-	}
-	n.setChainsyncIngressEligibility(e.ConnectionId, e.Eligible)
-}
-
-func shouldRecycleLocalTipPlateau(
-	now time.Time,
-	lastProgressAt time.Time,
-	localTipSlot uint64,
-	bestPeerTipSlot uint64,
-	lastRecycledAt *time.Time,
-	cooldown time.Duration,
-	threshold time.Duration,
-) bool {
-	if bestPeerTipSlot <= localTipSlot {
-		return false
-	}
-	if now.Sub(lastProgressAt) <= threshold {
-		return false
-	}
-	if lastRecycledAt != nil && now.Sub(*lastRecycledAt) < cooldown {
-		return false
-	}
-	return true
-}
-
-func (n *Node) processChainsyncRecyclerTick(
-	now time.Time,
-	localTipSlot uint64,
-	chainsyncCfg chainsync.Config,
-	recycleAt map[string]time.Time,
-	lastRecycled map[string]time.Time,
-	lastProgressSlot *uint64,
-	lastProgressAt *time.Time,
-	plateauRecoveryThreshold time.Duration,
-	grace time.Duration,
-	cooldown time.Duration,
-) {
-	if localTipSlot > *lastProgressSlot {
-		*lastProgressSlot = localTipSlot
-		*lastProgressAt = now
-	}
-	n.chainsyncState.CheckStalledClients()
-	trackedClients := n.chainsyncState.GetTrackedClients()
-	trackedByID := make(
-		map[string]chainsync.TrackedClient,
-		len(trackedClients),
-	)
-	for _, conn := range trackedClients {
-		connKey := conn.ConnId.String()
-		trackedByID[connKey] = conn
-		if conn.Status != chainsync.ClientStatusStalled {
-			delete(recycleAt, connKey)
-		}
-	}
-	// Prune expired cooldown entries so this map does
-	// not grow without bound over long runtimes.
-	for connKey, last := range lastRecycled {
-		if now.Sub(last) >= cooldown {
-			delete(lastRecycled, connKey)
-		}
-	}
-	// Safety net: if local tip has not moved for a long time
-	// while peers are ahead, recycle the selected chainsync
-	// connection even if it is not marked stalled.
-	if n.chainSelector != nil {
-		if bestPeer := n.chainSelector.GetBestPeer(); bestPeer != nil {
-			if bestPeerTip := n.chainSelector.GetPeerTip(*bestPeer); bestPeerTip != nil &&
-				bestPeerTip.Tip.Point.Slot > localTipSlot {
-				targetConn := n.chainsyncState.GetClientConnId()
-				if targetConn == nil {
-					targetCopy := *bestPeer
-					targetConn = &targetCopy
-				}
-				connKey := targetConn.String()
-				var lastRecycledAt *time.Time
-				if last, ok := lastRecycled[connKey]; ok {
-					lastCopy := last
-					lastRecycledAt = &lastCopy
-				}
-				if shouldRecycleLocalTipPlateau(
-					now,
-					*lastProgressAt,
-					localTipSlot,
-					bestPeerTip.Tip.Point.Slot,
-					lastRecycledAt,
-					cooldown,
-					plateauRecoveryThreshold,
-				) {
-					n.config.logger.Warn(
-						"local tip plateau detected, resyncing chainsync client",
-						"connection_id", connKey,
-						"local_tip_slot", localTipSlot,
-						"best_peer_tip_slot", bestPeerTip.Tip.Point.Slot,
-						"plateau_duration", now.Sub(*lastProgressAt),
-					)
-					n.eventBus.Publish(
-						event.ChainsyncResyncEventType,
-						event.NewEvent(
-							event.ChainsyncResyncEventType,
-							event.ChainsyncResyncEvent{
-								ConnectionId: *targetConn,
-								Reason:       "local_tip_plateau",
-							},
-						),
-					)
-					delete(recycleAt, connKey)
-					lastRecycled[connKey] = now
-					*lastProgressAt = now
-				}
-			}
-		}
-	}
-	for _, conn := range trackedClients {
-		if conn.Status != chainsync.ClientStatusStalled {
-			continue
-		}
-		connKey := conn.ConnId.String()
-		if _, exists := recycleAt[connKey]; !exists {
-			recycleAt[connKey] = now.Add(grace)
-			n.config.logger.Info(
-				"chainsync client stalled, scheduling guarded recycle",
-				"connection_id", connKey,
-				"stall_timeout", chainsyncCfg.StallTimeout,
-				"grace_period", grace,
-			)
-		}
-	}
-	for connKey, dueAt := range recycleAt {
-		if now.Before(dueAt) {
-			continue
-		}
-		tracked, ok := trackedByID[connKey]
-		if !ok || tracked.Status != chainsync.ClientStatusStalled {
-			delete(recycleAt, connKey)
-			continue
-		}
-		connId := tracked.ConnId
-		if last, ok := lastRecycled[connKey]; ok &&
-			now.Sub(last) < cooldown {
-			recycleAt[connKey] = now.Add(cooldown - now.Sub(last))
-			continue
-		}
-		active := n.chainsyncState.GetClientConnId()
-		if active == nil {
-			// If no active client is selected and this client
-			// is overdue + stalled, recycle to force a fresh
-			// connection attempt and avoid indefinite stalls.
-			n.config.logger.Warn(
-				"chainsync client stalled with no active selection, recycling connection",
-				"connection_id", connKey,
-				"stall_timeout", chainsyncCfg.StallTimeout,
-				"grace_period", grace,
-				"recycle_cooldown", cooldown,
-			)
-			n.eventBus.PublishAsync(
-				connmanager.ConnectionRecycleRequestedEventType,
-				event.NewEvent(
-					connmanager.ConnectionRecycleRequestedEventType,
-					connmanager.ConnectionRecycleRequestedEvent{
-						ConnectionId: connId,
-						ConnKey:      connKey,
-						Reason:       "stalled_connection_no_active_selection",
-					},
-				),
-			)
-			delete(recycleAt, connKey)
-			lastRecycled[connKey] = now
-			continue
-		}
-		if active.String() != connKey {
-			// Don't recycle non-primary stalled clients. Keep state clean.
-			n.eventBus.PublishAsync(
-				chainsync.ClientRemoveRequestedEventType,
-				event.NewEvent(
-					chainsync.ClientRemoveRequestedEventType,
-					chainsync.ClientRemoveRequestedEvent{
-						ConnId:  connId,
-						ConnKey: connKey,
-						Reason:  "stalled_non_primary_connection",
-					},
-				),
-			)
-			delete(recycleAt, connKey)
-			continue
-		}
-		n.config.logger.Warn(
-			"chainsync client stalled, recycling active connection",
-			"connection_id", connKey,
-			"stall_timeout", chainsyncCfg.StallTimeout,
-			"grace_period", grace,
-			"recycle_cooldown", cooldown,
-		)
-		n.eventBus.PublishAsync(
-			connmanager.ConnectionRecycleRequestedEventType,
-			event.NewEvent(
-				connmanager.ConnectionRecycleRequestedEventType,
-				connmanager.ConnectionRecycleRequestedEvent{
-					ConnectionId: connId,
-					ConnKey:      connKey,
-					Reason:       "stalled_active_connection",
-				},
-			),
-		)
-		delete(recycleAt, connKey)
-		lastRecycled[connKey] = now
-	}
-}
-
-func (n *Node) handleChainSwitchEvent(evt event.Event) {
-	e, ok := evt.Data.(chainselection.ChainSwitchEvent)
-	if !ok {
-		return
-	}
-	prevConn := "(none)"
-	if e.PreviousConnectionId.LocalAddr != nil &&
-		e.PreviousConnectionId.RemoteAddr != nil {
-		prevConn = e.PreviousConnectionId.String()
-	}
-	n.config.logger.Info(
-		"chain switch: updating active connection",
-		"previous_connection", prevConn,
-		"new_connection", e.NewConnectionId.String(),
-		"new_tip_block", e.NewTip.BlockNumber,
-		"new_tip_slot", e.NewTip.Point.Slot,
-	)
-	// Peer switches only change which already-running chainsync stream feeds
-	// the ledger. Restarting chainsync here re-enters FindIntersect and can
-	// race the protocol state machine under load.
-	n.chainsyncState.SetClientConnId(e.NewConnectionId)
 }
 
 func New(cfg Config) (*Node, error) {
@@ -1086,496 +800,4 @@ func (n *Node) Run(ctx context.Context) error {
 	// Wait for shutdown signal
 	<-n.ctx.Done()
 	return nil
-}
-
-func (n *Node) runStallCheckerTick(fn func()) {
-	defer func() {
-		if r := recover(); r != nil {
-			stack := debug.Stack()
-			n.config.logger.Error(
-				"panic in stall checker tick, continuing",
-				"panic", r,
-				"stack", string(stack),
-			)
-		}
-	}()
-	fn()
-}
-
-func (n *Node) runStallCheckerLoop(fn func()) (recovered bool) {
-	defer func() {
-		if r := recover(); r != nil {
-			recovered = true
-			stack := debug.Stack()
-			n.config.logger.Error(
-				"panic in stall checker goroutine",
-				"panic", r,
-				"stack", string(stack),
-			)
-		}
-	}()
-	fn()
-	return false
-}
-
-func (n *Node) Stop() error {
-	var err error
-	n.shutdownOnce.Do(func() {
-		err = n.shutdown()
-	})
-	return err
-}
-
-func (n *Node) shutdown() error {
-	shutdownStart := time.Now()
-	// Create shutdown context with timeout (default 30s if not configured)
-	shutdownTimeout := 30 * time.Second
-	if n.config.shutdownTimeout > 0 {
-		shutdownTimeout = n.config.shutdownTimeout
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), shutdownTimeout)
-	defer cancel()
-	if n.cancel != nil {
-		n.cancel()
-	}
-
-	var err error
-
-	n.config.logger.Info(
-		"starting graceful shutdown",
-		"timeout", shutdownTimeout,
-	)
-
-	// Phase 1: Stop accepting new work
-	n.config.logger.Info("shutdown phase 1: stopping new work")
-
-	// Stop block forger first to prevent new blocks
-	if n.blockForger != nil {
-		n.blockForger.Stop()
-	}
-
-	// Stop leader election to clean up resources
-	if n.leaderElection != nil {
-		if stopErr := n.leaderElection.Stop(); stopErr != nil {
-			err = errors.Join(
-				err,
-				fmt.Errorf("leader election shutdown: %w", stopErr),
-			)
-		}
-	}
-
-	if n.chainSelector != nil {
-		n.chainSelector.Stop()
-	}
-
-	if n.peerGov != nil {
-		n.peerGov.Stop()
-	}
-
-	if n.snapshotMgr != nil {
-		if stopErr := n.snapshotMgr.Stop(); stopErr != nil {
-			err = errors.Join(
-				err,
-				fmt.Errorf("snapshot manager shutdown: %w", stopErr),
-			)
-		}
-	}
-
-	if n.utxorpc != nil {
-		if stopErr := n.utxorpc.Stop(ctx); stopErr != nil {
-			err = errors.Join(err, fmt.Errorf("utxorpc shutdown: %w", stopErr))
-		}
-	}
-
-	if n.bark != nil {
-		if stopErr := n.bark.Stop(ctx); stopErr != nil {
-			err = errors.Join(err, fmt.Errorf("bark shutdown: %w", stopErr))
-		}
-	}
-
-	if n.barkPruner != nil {
-		if stopErr := n.barkPruner.Stop(ctx); stopErr != nil {
-			err = errors.Join(
-				err,
-				fmt.Errorf("bark pruner shutdown: %w", stopErr))
-		}
-	}
-
-	if n.blockfrostAPI != nil {
-		if stopErr := n.blockfrostAPI.Stop(ctx); stopErr != nil {
-			err = errors.Join(
-				err,
-				fmt.Errorf("blockfrost API shutdown: %w", stopErr),
-			)
-		}
-	}
-
-	if n.meshAPI != nil {
-		if stopErr := n.meshAPI.Stop(ctx); stopErr != nil {
-			err = errors.Join(
-				err,
-				fmt.Errorf("mesh API shutdown: %w", stopErr),
-			)
-		}
-	}
-
-	n.config.logger.Info(
-		"shutdown phase 1 complete",
-		"elapsed", time.Since(shutdownStart).Round(time.Millisecond),
-	)
-
-	// Phase 2: Drain and close connections
-	n.config.logger.Info("shutdown phase 2: draining connections")
-	phase2Start := time.Now()
-
-	if n.mempool != nil {
-		if stopErr := n.mempool.Stop(ctx); stopErr != nil {
-			err = errors.Join(err, fmt.Errorf("mempool shutdown: %w", stopErr))
-		}
-	}
-
-	if n.connManager != nil {
-		if stopErr := n.connManager.Stop(ctx); stopErr != nil {
-			err = errors.Join(
-				err,
-				fmt.Errorf("connection manager shutdown: %w", stopErr),
-			)
-		}
-	}
-
-	n.config.logger.Info(
-		"shutdown phase 2 complete",
-		"elapsed", time.Since(phase2Start).Round(time.Millisecond),
-	)
-
-	// Phase 3: Flush state and close database
-	n.config.logger.Info("shutdown phase 3: flushing state")
-	phase3Start := time.Now()
-
-	if n.ledgerState != nil {
-		n.config.logger.Info("closing ledger state")
-		t := time.Now()
-		if closeErr := n.ledgerState.Close(); closeErr != nil {
-			err = errors.Join(
-				err,
-				fmt.Errorf("ledger state close: %w", closeErr),
-			)
-		}
-		n.config.logger.Info(
-			"ledger state closed",
-			"elapsed", time.Since(t).Round(time.Millisecond),
-		)
-	}
-
-	if n.db != nil {
-		n.config.logger.Info("closing database")
-		t := time.Now()
-		if closeErr := n.db.Close(); closeErr != nil {
-			err = errors.Join(
-				err,
-				fmt.Errorf("database close: %w", closeErr),
-			)
-		}
-		n.config.logger.Info(
-			"database closed",
-			"elapsed", time.Since(t).Round(time.Millisecond),
-		)
-	}
-
-	n.config.logger.Info(
-		"shutdown phase 3 complete",
-		"elapsed", time.Since(phase3Start).Round(time.Millisecond),
-	)
-
-	// Phase 4: Cleanup resources
-	n.config.logger.Info("shutdown phase 4: cleanup resources")
-
-	// Call registered shutdown functions
-	for _, fn := range n.shutdownFuncs {
-		if fnErr := fn(ctx); fnErr != nil {
-			err = errors.Join(err, fmt.Errorf("shutdown function: %w", fnErr))
-		}
-	}
-	n.shutdownFuncs = nil
-
-	if n.eventBus != nil {
-		n.eventBus.Stop()
-	}
-
-	n.config.logger.Info(
-		"graceful shutdown complete",
-		"total_elapsed", time.Since(shutdownStart).Round(time.Millisecond),
-	)
-	return err
-}
-
-func (n *Node) validateBlockProducerStartup() (*forging.PoolCredentials, error) {
-	creds := forging.NewPoolCredentials()
-	if err := creds.LoadFromFiles(
-		n.config.shelleyVRFKey,
-		n.config.shelleyKESKey,
-		n.config.shelleyOperationalCertificate,
-	); err != nil {
-		return nil, fmt.Errorf("load pool credentials: %w", err)
-	}
-	if err := creds.ValidateOpCert(); err != nil {
-		return nil, fmt.Errorf("validate operational certificate: %w", err)
-	}
-	n.config.logger.Info(
-		"block producer credentials validated",
-		"pool_id", creds.GetPoolID().String(),
-		"opcert_expiry_period", creds.OpCertExpiryPeriod(),
-	)
-	return creds, nil
-}
-
-// initBlockForger initializes the block forger for production mode.
-// This requires VRF, KES, and OpCert key files to be configured.
-func (n *Node) initBlockForger(
-	ctx context.Context,
-	creds *forging.PoolCredentials,
-) error {
-	if creds == nil {
-		return errors.New("nil pool credentials")
-	}
-	// Create mempool adapter for the forging package
-	mempoolAdapter := &mempoolAdapter{mempool: n.mempool}
-
-	// Create epoch nonce adapter for the builder
-	epochNonceAdapter := &epochNonceAdapter{ledgerState: n.ledgerState}
-
-	// Create block builder
-	builder, err := forging.NewDefaultBlockBuilder(forging.BlockBuilderConfig{
-		Logger:          n.config.logger,
-		Mempool:         mempoolAdapter,
-		PParamsProvider: n.ledgerState,
-		ChainTip:        n.chainManager.PrimaryChain(),
-		EpochNonce:      epochNonceAdapter,
-		Credentials:     creds,
-		TxValidator:     n.ledgerState,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to create block builder: %w", err)
-	}
-
-	// Create block broadcaster (uses the chain manager and event bus)
-	broadcaster := &blockBroadcaster{
-		chain:    n.chainManager.PrimaryChain(),
-		eventBus: n.eventBus,
-		logger:   n.config.logger,
-	}
-
-	// Create the leader election component
-	// Convert pool ID from PoolId to PoolKeyHash (both are [28]byte)
-	poolID := creds.GetPoolID()
-	var poolKeyHash lcommon.PoolKeyHash
-	copy(poolKeyHash[:], poolID[:])
-
-	// Create adapters for the providers that leader.Election needs
-	stakeProvider := &stakeDistributionAdapter{ledgerState: n.ledgerState}
-	epochProvider := &epochInfoAdapter{ledgerState: n.ledgerState}
-
-	// Get VRF secret key from credentials
-	vrfSKey := creds.GetVRFSKey()
-
-	// Create leader election with real stake distribution
-	election := leader.NewElection(
-		poolKeyHash,
-		vrfSKey,
-		stakeProvider,
-		epochProvider,
-		n.eventBus,
-		n.config.logger,
-	)
-	if n.db != nil {
-		if scheduleStore := leader.NewSyncStateScheduleStore(
-			n.db.Metadata(),
-		); scheduleStore != nil {
-			election.SetScheduleStore(scheduleStore)
-		}
-	}
-
-	// Start leader election (subscribes to epoch transitions)
-	if err := election.Start(ctx); err != nil {
-		return fmt.Errorf("failed to start leader election: %w", err)
-	}
-
-	// Store election for cleanup during shutdown
-	n.leaderElection = election
-
-	// Create slot clock adapter for the forger
-	slotClock := &slotClockAdapter{ledgerState: n.ledgerState}
-
-	// Create the block forger with the real leader election
-	forger, err := forging.NewBlockForger(forging.ForgerConfig{
-		Mode:                        forging.ModeProduction,
-		Logger:                      n.config.logger,
-		Credentials:                 creds,
-		LeaderChecker:               election,
-		BlockBuilder:                builder,
-		BlockBroadcaster:            broadcaster,
-		SlotClock:                   slotClock,
-		ForgeSyncToleranceSlots:     n.config.forgeSyncToleranceSlots,
-		ForgeStaleGapThresholdSlots: n.config.forgeStaleGapThresholdSlots,
-		PromRegistry:                n.config.promRegistry,
-	})
-	if err != nil {
-		// Stop election to prevent goroutine leak
-		_ = election.Stop()
-		return fmt.Errorf("failed to create block forger: %w", err)
-	}
-
-	// Start the forger with the passed context
-	if err := forger.Start(ctx); err != nil {
-		// Stop election to prevent goroutine leak
-		_ = election.Stop()
-		return fmt.Errorf("failed to start block forger: %w", err)
-	}
-
-	n.blockForger = forger
-	n.config.logger.Info(
-		"block forger started in production mode with leader election",
-		"pool_id", poolID.String(),
-	)
-
-	return nil
-}
-
-// mempoolAdapter adapts the mempool.Mempool to forging.MempoolProvider.
-type mempoolAdapter struct {
-	mempool *mempool.Mempool
-}
-
-func (a *mempoolAdapter) Transactions() []forging.MempoolTransaction {
-	txs := a.mempool.Transactions()
-	result := make([]forging.MempoolTransaction, len(txs))
-	for i, tx := range txs {
-		result[i] = forging.MempoolTransaction{
-			Hash: tx.Hash,
-			Cbor: tx.Cbor,
-			Type: tx.Type,
-		}
-	}
-	return result
-}
-
-// blockBroadcaster implements forging.BlockBroadcaster using the chain manager.
-type blockBroadcaster struct {
-	chain    *chain.Chain
-	eventBus *event.EventBus
-	logger   *slog.Logger
-}
-
-func (b *blockBroadcaster) AddBlock(
-	block gledger.Block,
-	_ []byte,
-) error {
-	// Add block to the chain (CBOR is stored internally by the block).
-	// addBlockInternal notifies iterators after storing the block.
-	if err := b.chain.AddBlock(block, nil); err != nil {
-		return fmt.Errorf("failed to add block to chain: %w", err)
-	}
-
-	b.logger.Info(
-		"block added to chain",
-		"slot", block.SlotNumber(),
-		"hash", block.Hash(),
-		"block_number", block.BlockNumber(),
-	)
-
-	return nil
-}
-
-// stakeDistributionAdapter adapts ledger.LedgerState to leader.StakeDistributionProvider.
-// It queries the metadata store directly with a nil transaction so the SQLite
-// read pool is used, avoiding contention with block-processing write locks.
-type stakeDistributionAdapter struct {
-	ledgerState *ledger.LedgerState
-}
-
-func (a *stakeDistributionAdapter) GetPoolStake(
-	epoch uint64,
-	poolKeyHash []byte,
-) (uint64, error) {
-	snapshot, err := a.ledgerState.Database().Metadata().GetPoolStakeSnapshot(
-		epoch, "mark", poolKeyHash, nil,
-	)
-	if err != nil {
-		return 0, err
-	}
-	if snapshot == nil {
-		return 0, nil
-	}
-	return uint64(snapshot.TotalStake), nil
-}
-
-func (a *stakeDistributionAdapter) GetTotalActiveStake(
-	epoch uint64,
-) (uint64, error) {
-	return a.ledgerState.Database().Metadata().GetTotalActiveStake(
-		epoch, "mark", nil,
-	)
-}
-
-// epochInfoAdapter adapts ledger.LedgerState to leader.EpochInfoProvider.
-type epochInfoAdapter struct {
-	ledgerState *ledger.LedgerState
-}
-
-func (a *epochInfoAdapter) CurrentEpoch() uint64 {
-	return a.ledgerState.CurrentEpoch()
-}
-
-func (a *epochInfoAdapter) EpochNonce(epoch uint64) []byte {
-	return a.ledgerState.EpochNonce(epoch)
-}
-
-func (a *epochInfoAdapter) NextEpochNonceReadyEpoch() (uint64, bool) {
-	return a.ledgerState.NextEpochNonceReadyEpoch()
-}
-
-func (a *epochInfoAdapter) SlotsPerEpoch() uint64 {
-	return a.ledgerState.SlotsPerEpoch()
-}
-
-func (a *epochInfoAdapter) ActiveSlotCoeff() float64 {
-	return a.ledgerState.ActiveSlotCoeff()
-}
-
-// slotClockAdapter adapts ledger.LedgerState to forging.SlotClockProvider.
-type slotClockAdapter struct {
-	ledgerState *ledger.LedgerState
-}
-
-func (a *slotClockAdapter) CurrentSlot() (uint64, error) {
-	return a.ledgerState.CurrentSlot()
-}
-
-func (a *slotClockAdapter) SlotsPerKESPeriod() uint64 {
-	return a.ledgerState.SlotsPerKESPeriod()
-}
-
-func (a *slotClockAdapter) ChainTipSlot() uint64 {
-	return a.ledgerState.ChainTipSlot()
-}
-
-func (a *slotClockAdapter) NextSlotTime() (time.Time, error) {
-	return a.ledgerState.NextSlotTime()
-}
-
-func (a *slotClockAdapter) UpstreamTipSlot() uint64 {
-	return a.ledgerState.UpstreamTipSlot()
-}
-
-// epochNonceAdapter adapts ledger.LedgerState to forging.EpochNonceProvider.
-type epochNonceAdapter struct {
-	ledgerState *ledger.LedgerState
-}
-
-func (a *epochNonceAdapter) CurrentEpoch() uint64 {
-	return a.ledgerState.CurrentEpoch()
-}
-
-func (a *epochNonceAdapter) EpochNonce(epoch uint64) []byte {
-	return a.ledgerState.EpochNonce(epoch)
 }
